tighten up object lookups
over time the db has accumulated a fair number of subjects and
employers with duplicate names, which causes problems for
views that use a slugified version of the name.

this commit tightens up the lookups to use the freebase id
and also includes a new command line utility to help diagnose
and correct these duplicates.

refs #38